startup_proj/requirements/develop.txt
# Development dependencies
-r _base.txt

django-debug-toolbar
django-sass-processor

Werkzeug  # Used to run django-extensions runserver_plus
